Create an organization and its departments from CSV files
- Add `createOrganizationAndDepartments` method to create an organization and its departments from CSV files based on the given organization name, returning a reference to the organization.
  - Parse organization info from `organization_info.csv`.
  - Iterate through CSV files to extract department data, current manager, employees, former managers, and former employees.

+++ src/main/java/com/example/salaryproject/AddEmployeePageController.java
@@ -127,8 +127,8 @@ public class AddEmployeePageController {
             messageLabel.setStyle("-fx-text-fill: red;");
         } else {
             try {
-                long nationalId = Long.parseLong(nationalIdText);
-                long phoneNumber = Long.parseLong(phoneNumberText);
+                int nationalId = Integer.parseInt(nationalIdText);
+                int phoneNumber = Integer.parseInt(phoneNumberText);
 
                 if (selectedDepartment == null) {
                     messageLabel.setText("No department selected!");
